refactor: move content of output.text.utils to core.text_utils
