+++ ggshield/core/text_utils.py
@@ -1,6 +1,6 @@
 import sys
 from enum import Enum, auto
-from typing import Any, Dict, NamedTuple, Optional, Union
+from typing import Any, Dict, List, NamedTuple, Optional, Union
 
 import click
 from rich.console import Console
@@ -165,3 +165,17 @@ def create_progress_bar(doc_type: str) -> Progress:
         TimeRemainingColumn(),
         console=Console(file=sys.stderr),
     )
+
+
+def get_padding(lines: List[Line]) -> int:
+    """Return the number of digit of the maximum line number."""
+    # value can be None
+    return max(len(str(lines[-1].pre_index or 0)), len(str(lines[-1].post_index or 0)))
+
+
+def get_offset(padding: int, is_patch: bool = False) -> int:
+    """Return the offset due to the line display."""
+    if is_patch:
+        return len(LINE_DISPLAY["patch"].format("0" * padding, "0" * padding))
+
+    return len(LINE_DISPLAY["file"].format("0" * padding))
